Really fix config files location
While making the previous patch to fix the config files location, I added
the sysconfdir=/etc flag both to make and make install and tested that it
worked fine.

I then tested without sysconfdir=/etc flag for make and the resulting
binary also worked, so I removed it.

However I was building with makepkg, so it kept the result of the previous
compilation and just re-run make.

As the target didn't change, the make process probably didn't recompile
anything, and at testing everything worked fine, while in reality the
PKGBUILD cound't produce the proper fix.

This is because the configuration file location are passed to the C
source code as compilation flags through this like in the Makefile:
    CONFCPPFLAGS = -DSYSCONFDIR='"$(sysconfdir)"' -DREL_SYSCONFDIR='"$(rel_sysconfdir)"'

When rebuilding from scratch, without passing sysconfdir=/etc to make, we
have:
    $ strings $(which imx_usb) | grep "imx-loader\.d"
    /usr/etc/imx-loader.d/
    ../etc/imx-loader.d

And when rebuilding from scratch and passing sysconfdir=/etc to make, we
finally have:
    $ strings $(which imx_usb) | grep "imx-loader\.d"
    /etc/imx-loader.d/
    ../../etc/imx-loader.d

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>