Initialize a package to build the GRUB for the Xen platform
Since the time pvgrub2 was adopted by the GRUB [1], the AUR has not had
a package that built GRUB2 for the Xen platform. I thought about
contributing a patch to the `grub-git` AUR package, but since it would
double their build targets (and compile time) for a purpose that was
probably not important or useful for most of the users of that package
(and since this also seemed like too much for this package to take on as
well), I thought it might be better to fork their project and create a
new package that would address all the GRUB-ish needs of Xen.

The intent is that this would follow `grub-git`, rebasing it on that
package as needed. Also some of these patches might be useful upstream,
either for the `grub-git` AUR package or for the GRUB project itself.

[1]: https://lists.xen.org/archives/html/xen-devel/2013-11/msg01216.html