Temporarily disable glslang. Fix svt-hevc doc patch.
Upstream ffmpeg now requires[1] a glslang version that is not
yet on the official repositories. glslang-git from the AUR can
be used, but the package needs a lot of improvements (see
comments on the package page). Temporarily disable it until we
have a supported glslang version on the official repositories
or until glslang-git is improved.

The patch for svt-hevc doc currently fails to apply becasue
there was a split on upstream file doc/general.texi[2]. Now
it's splitted into doc/general_contents.texi. A local modification
of the patch does the job.

Upstream svt-hevc updated[3] their patch to work with the current
ffmpeg git master, so we do not need the custom patch anymore.
Now using the svt-hevc patch directly from upstream.

Also on this commit: refresh the vmaf and cuda11 patches.

References
----------
[1] https://git.videolan.org/?p=ffmpeg.git;a=commitdiff;h=98ea1a662ee172961feaa374fe8d26078838d250
[2] https://git.videolan.org/?p=ffmpeg.git;a=commitdiff;h=6accb7718aa4c9f5ca8f3280a1ed4e09587a627f
[3] https://github.com/OpenVisualCloud/SVT-HEVC/commit/ead6fdf7c9ff84511b42fc1658c1654b84d83e4b