patch 'os.h' to allow slurm to build
Following the removal of the 'sysctl' syscall in Linux 5.5, the GNU libc has
removed the sys/sysctl.h header and associated function in version 2.32 (see the
release notes: https://sourceware.org/pipermail/libc-announce/2020/000029.html).

The 'os.h' file includes this removed header. This commit introduces the
'sysctl.patch' file, which removes the inclusion of the the 'sys/sysctl.h' file.
The software seems to build and run fine without it.