Major refactoring
The key guideline that I followed here is to avoid patches as much as
possible, in order to keep this package simple and clean.

So I deciced to not depend on ninja and python2, and instead just use
those that come with depot_tools.

Also, I think that most of the python2 issues have been resolved
upstream by now, and there is no need to force scripts to run using
python2 anymore.

Finally, I did many small cleanups in PKGBUILD, and added bash
completion for gclient and git cl.