Default to Electron 8 dependency
Previously Electron 7 was in [community] so we preferred that over
needing to install 8 from the AUR, but now 7 has gone away too. Since
users are forced to use an AUR bin package, we might as well default to
8 again.

This dependency is provided by `electron8-bin` as of this time. If
anybody wants to provide a source package `electron8` that would be
awesome, and that would become a viable option for fulfilling this
dependency.