Bump pkgrel for config hash change
It seems that sometimes the Clear Linux version is updated to 33760,
for example, but the config hash is only available for 33750... this
is related to the issues I had updating the linux-clear-bin package
recently where they bump the Clear version and ship a new manifest,
but the kernel version itself isn't changed. In this case it is only
the config for the compiled kernel (ie, /proc/config), but ugh.