Update to v1.5
Also Drop Makefile patch which allows building with system version of
libraries, as it is now part of the upstream release.

Also add 'libappindicator-gtk3' as build dependency, as Droidcam will
now register application indicator while running. The indicator also
requires an PNG icon to be loaded from /opt/droidcam-icon.png, so we
place a copy there.

Signed-off-by: Mateusz Gozdek <mgozdekof@gmail.com>