Update download link and navigaton commands
byuu's original site is no longer available. :(

In its place is a Markdown-based rendition of the core site content.
The file server no longer seems to be available, which means the
original dowloads for amethyst are all but gone. The site doesn't
appear in the Wayback Machine either. I didn't try looking at Google's
Web Cache or similar services, but those are likely temporary anyway.

Luckily, I had a copy of amethyst in a Git repository, which I have
made public and pointed this package to.

This repository contains license files, so the LICENSE file in this
package has been removed and all references in the PKGBUILD have also
been removed. I have also removed references to the .install file as,
according to https://wiki.archlinux.org/index.php/PKGBUILD#source this
should not be listed.

With a new way of downloading the code comes new ways of navigating
and patching said code. Namely the previous tarball did not have a
containing directory, whereas this one does. This means adjusting the
'cd' commands and patching the top-level directory (-p0) instead of
one directory down (-p1).

We also need to add the build directories manually. The original
tarball contained these as empty directories, but these could not be
pushed to the Git repository.