Clean up the signature verification
Winfried Brügmann changed his key and revoked the old one, so we had
to switch to the new one. In this case, the signature files got changed
too (reason for the changed sha512 sums). If you got issues with the
signature process, make sure you import (and trust) the new key
3F0CC709ECF91C5CC0BE9E601D295C19C9C06AF6.