upgpkg: calibre-git 5.3.0.r35.g40e5b7124f-1
Stop depending on ancient libusbx provides, because libusb is once again
the preferred implementation... for 6 years now, so depending on the
provides is bleurgh. And in [testing], it's no longer provided.

Also drop dependency on chmlib now this is only used via pychm, and
switch one holdout command invocation from python3 -> python.