use the git hash directly + rebuild with GCC 10
Also make it clear that the package us using the latest git tag from github, not from astian.org.

The new webpage for Midori is https://astian.org/en/midori-browser/, but they have no links to any
development activity and seem to have departed from the previous versioning scheme.

The Midori project was recently merged with the Astian Foundation. Once there is a link to
a source repository there, I'm planning to switch over the package URL to that web page.