Initial Commit
PKGBUILD and .SRCINFO added.

The packaged xsuspender.desktop file contains the line 'Hidden=true' which stops it from functioning. It may be desirable to remove it, and this has been mentioned upstream. The result is user intervention is required for xsuspender to run on DE start.

Configuration must also be made in ~/.config/xsuspender.conf (example provided), but since this is user configuration and varies wildly for different systems/users it is outside the scope of a PKGBUILD.