Remove broken dependency on libusbx
The most recent package release of libusb, 1.0.23-3, removed the
'libusbx=1.0.23' provides directive. That broke the 'libusbx>=1.0.8'
dependency in this package.

It actually turns out that libftd2xx has been statically linked to its
own internal copy of libusb for a VERY LONG time, judging by the
release-notes.txt file: the notes for version 0.4.0 (!) specifically
mention switching to "static linkage to libusb to avoid installation
issues."

I checked both the dynamic libftd2xx.so file and the static libftd2xx.a
file and confirmed that both library variants do not have any external
dependency on libusb; all of the required libusb symbols are contained
within the libftd2xx library files themselves.

Incidentally, the libusbx dependency in the PKGBUILD, removed in this
commit, suggested a minimum required libusb version of 1.0.8. Meanwhile,
the included copy of libusb is version 1.0.20. So it's entirely possible
that, even IF the premise that an external libusb was needed was
correct, that the version asked for as a dependency may have not even
been sufficient.