Bump to version 1.5.0
This new release now includes the IEEE VHDL libraries, so we don't
need to download and copy them around by hand. The patches to use
newer LLVM versions are not needed anymore. Lastly, now we pass a
few --enable-foo and --with-bar options to the configure script to
enable additional features.