Update to 1.6
For now without "official" support for v4l2loopback-dkms package, though
it can be installed and used for testing. This is because I'm not sure
yet how to deal with it as optional dependency and also while testing
myself, I found it difficult to set the desired resolution, so in my
opinion v4l2loopback-dc-dkms still provides a bit better UX than the
other module.

Signed-off-by: Mateusz Gozdek <mgozdekof@gmail.com>