Fix tests failing for the wrong reason
As pytest.ini does a lot of things we don't require for testing the package
such as pep8 and coverage remove it to allow us to run only the tests. This
also stops the downloading of pip packages during testing which is unwanted.

Additionally rename ${pkgbase} to ${pkgname} as ${pkgbase} is not set.