Add GraalWasm to test.sh script
Needs Emscripten to compile the C file, which apparently puts the emcc
binary in an unusual location, so add that to the PATH as well. I also
considered adding the .wasm file to the Git repository directly, but it
seems the format changes occasionally – a file I used to test earlier
GraalWasm releases didn’t work with 20.3.0 anymore – so let’s instead
have the test script compile the source code from scratch, and require
the testing user to have emscripten installed.