Various Improvements
Various improvements from opendds. Also use pkgver function. The problem
with this is that it requires a full git clone, which I tried to avoid
before with the git sources that don't support --depth=1. Also a shallow
copy doesn't have the tags that the pkgver function needs. Thankfully I
only have to use a full git clone on OpenDDS and not ACE_TAO, which
takes a lot longer than ACE_TAO to fully clone.