Update Vampire to 4.5
Vampire now supports building with CMake, so use that instead of the
handwritten Makefile.

The separate patch is used for two reasons. First, Vampire includes the
current Git commit information in its --version output (commit hash and
timestamp). When building this package from AUR, this would read the
commit metadata from AUR and not the Vampire repository. Second, Vampire
uses the __FILE__ macro even in release builds, which means that
references to the build directory end up in the binary. To avoid this,
a GCC- and Clang-specific compiler option is added to ensure that the
file references are relative to Vampire’s source directory.