Tolerate missing pandoc
Pandoc is a heavy package to require – written in Haskell, it can pull
in a lot of packages that wouldn’t otherwise be needed: when I
uninstalled it locally, over 200 MiB were removed. We can’t mark the
dependency as optional (there’s no optmakedepends/makeoptdepends), but
we can make the build script tolerate a missing pandoc a bit better, and
users who really want to avoid pandoc can then run makepkg with -d to
ignore the dependency at their own risk.