fix fallback directory detection
firefox (probably due to sandboxing) results in /proc-based cwds. This
can be annoying. So swaycwd will now detect if the cwd still is in /proc
and return the fallback dir.

The Fallbackdir (previously was hardcoded to /home/julian) can now be passed via a
command line arg. And since commandline args are parsed now, we can also
print a helptext as well :)