Update pkgver in PKGBUILD for the pkgrel bump to take effect
pkgver is formatted dynamically, so the installed package had version
r17.e2cd097-1, but the updates are checked against the static version in
the PKGBUILD, which got bumped only to r16.931e2c7-2 in the previous
commit. Fix this by also updating the pkgver. Note that pkgrel still has
to be 2 to trigger an update and rebuild for Python 3.9, because the
installed version is r17.e2cd097-1.