Initial package release
Espresso was already released with GraalVM 21.0.0 and 21.0.0.2, but
those releases did not include a license file, so I did not feel
comfortable publishing them. With 21.1.0, we have a license file, and it
appears to be a standard GPLv2 license, so that’s what we put in the
license field as well.