Use b2sums instead of md5sums
We all know MD5 has been vulnerable to a myriad of attacks for a long
time. Although this package most likely won't be "attacked," it is
reasonable to ditch MD5 and migrate to a known secure hash algorithm
since MD5 has proven it can no longer accomplish its job effectively.

If for no other reason, BLAKE2b is faster than MD5 to compute, so why
not switch?