Package CasparCG client from their Git repository
Using the included binary libs fails for libqatemcontrol:

    /usr/bin/ld: [...]/src/Client/src/Shell/../../lib/qatemcontrol/lib/linux/release//libqatemcontrol.so:
    undefined reference to `operator delete(void*, unsigned long)@Qt_5'

And also for gpio-client:

    /usr/bin/ld: [...]src/Client/src/Shell/../../lib/gpio-client/lib/linux//libgpio-client.so:
    undefined reference to `boost::future_category()'
    undefined reference to `boost::this_thread::hiden::sleep_for(timespec const&)'
    ...

Using the provided `build-system-linux.sh` script fails as well:

    Running qmake...
    ***Unknown option -qt=5

This package uses the `build-linux.sh` script but patches it to use
system libs where necessary to build and link a working executable.

The gpio-client lib is built as part of this package since it's tightly
coupled with the CasparCG Client and not available as a standalone
library. Therefor it was decided to not package it separately but
rebuild it here.