v1.0.0
Upgrade to the upstream v1.0.0 release, and clean up the PKGBUILD.

Note that this release is not compatible with previous versions of
fscryptctl.  If you need to unlock directories that were encrypted by a
previous version of fscryptctl, you'll need to use that older version.
See the installed README.md file for details.