Add missing licenses
While the different licenses used by the individual scripts do not allow
for the entire package to be licensed under a unified set of terms, I
think it is reasonable to interpret this as a collection of independent
(and independently licensed) scripts.

Additionally, when packaged on the end user's machine (as is normally
the case with AUR PKGBUILDs), there is no distribution of the package as
whole.