set custom name for source archive
The source archive filename does not contain any version numbers and
generally never changes. This means that AUR helpers see the old file
and attempt to use it without downloading the new one, but the checksum
fails. Disabling the checksum is not a solution because then the AUR
helpers would just use the old chached version.

With this commit each new source file is named with the version number
so they do not conflict.