Remove preloading via `go mod download`
There seems to be no way of providing `-modcacherw` or something
equivalent to `go mod download`, resulting in the option to be ignored
later in `go build`. This causes the mod files in src/ to be read-only,
which is not ideal. Since the benefits of preloading are contested
anyway, we default to the simplest solution and remove it for now.