New upstream option: --enable-librist. Re-enable glslang. Add lto.
Upstream now supports librist[1].

glslang support was temporarily disabled previously[2] and now it
can be re-enabled since glslang on the official repositories was
updated.

Repository ffmpeg added LTO, so we add it here too. RFC about LTO
on Arch Linux is being discussed and is on the way to be approved.

Also on this commit:
 - update svt-vp9 patch
 - refresh cuda11 patch
[1] https://git.videolan.org/?p=ffmpeg.git;a=commitdiff;h=4098f809d605fc540870883a9f91880e13ce1d6b
[2] https://aur.archlinux.org/cgit/aur.git/commit/?h=ffmpeg-full-git&id=bd252d99089862a9c74d43a88837d243c52de38e
[3] https://github.com/archlinux/svntogit-packages/commit/ceff68a52e22bfedf1f8e75228a608b7ddab3d15