Initial commit
GraalVM 21.1.0 is the first release that comes with an (experimental)
JDK16 version. The package files are mostly based on jdk11-graalvm-bin,
but I’ve had to remove parts of the test.sh script, since not all
components are available for JDK16 yet. (The original script is
version-independent, so once all components have JDK16 versions, I can
just copy the JDK11 test.sh again and discard this version.)