Initial commit
With GraalVM 21.1.0, the Node.js component is no longer included in the
base GraalVM download, but should be installed separately. Consequently,
we expose it as a separate AUR package as well.

The release artifact does not include a license file, so we download one
from GitHub. (It matches the GraalJS/Node.js section of the GraalVM
21.0.0.2 LICENSE.txt, at least.)