Fix inhibition of processor-specific optimizations
The build system tries to be smart and add -march=native by default,
but this is not what we want. Setting ARCH_OPT to an empty string
prevents an additional -march= from being generated, while still
allowing for detection of CPU features available for the architecture
specified by makepkg.conf.