build: Put vlc-luajit as dependency and 27 prepare
The official VLC package will make OBS crash when a VLC source is used.
The issue is that VLC and OBS are compiled with different lua version.
So I also created vlc-luajit, a VLC package compiled with the same lua as OBS.
But to make people unable to install VLC official package with obs-studio-tytan652.
I decided to make vlc-luajit a dependency of OBS rather than an optional one.