util-linux-aes: update to 2.37 (UNOFFICIAL)
This is a necessary update for util-linux to avoid breakage of systems.
Sadly there is no new upstream release, so I needed to adapt the patch myself.
Therefore this notice:
This patch isn't the official one. I could have done anything with it.
And it is incomplete. Due to upstream changing their man-page source, this patch doesn't include the changes for the man-pages. There is currently no documentation.
Source code changes were fairly easy to adapt. BUT they're untested and it is possible something broke in a subtle way.