Fix version info for libwlroots.so and wlroots
The provides version info for libwlroots.so is wrong and the version
info for wlroots is missing completely.

Version info for provides for .so files seems to be automatically
provided by makepkg (taken from the SONAME of the library). This is
also what the community package for wlroots does (wlroots currently
has libwlroots.so=9-64, wlroots-git would have libwlroots.so=10-64
with the automatic versioning)

Signed-off-by: Antonin Décimo <antonin.decimo@gmail.com>