Bump to 1.17, BuildTools 131; hard-code openjdk
Since it is now requiring a _specific_ version instead of a range, it
seems wrong to just use the default. Since there's only one, just
hard-code the path to the java to use.

That'll be annoying in September when JDK 17 comes out with the
instantaneous death of JDK 16, but there don't seem to be any other good
solutions since there is literally no overlap between the two versions
being available. At least JDK 17 will be LTS so there will be a respite.