Be honest about using upstream binary, fix chroot build
This package has never properly built from source, it always used the
binary wheel for Skia. Previous to this commit the machinery for that
was just hidden one layer down. The pathops repo was being "built", but
the build process was just using pip behind the scenes to fetch the
wheel.

*Actually* building this from source requires Python2 and ninja and all
sorts of song and dance that I wasn't able to untangle in the time
required. If/when anybody works on this they can start with the previous
commit as a model, but it needs lots of work. At least this version is
honest and does what it says on the tin.