Update to 21.2.0
GraalPython and TruffleRuby come with JDK16 versions now, so the test.sh
script is closer to the JDK11 version. However, Espresso still doesn’t
have a JDK16 version (reportedly there will be a JDK17 version once the
upcoming Java 17, an LTS version, is released), so the Espresso block is
still missing.