Use Node.js launcher in test.sh
The Node.js component is no longer included in this package, making it
separate from the JS language (which is still included), so we should
make sure to test it in test.sh. The simplest way to do that is to
replace R with Node as the host language for the polyglot test.