Use Node.js launcher in test.sh
The Node.js component is not included in this package, making it
separate from the JS language (which is included), so we should make
sure to test it in test.sh. The simplest way to do that is to replace R
with Node as the host language for the polyglot test.