remove out-date dependencies
The library files packed in the .deb is made before 2019. These libraries are
not compatible with arch linux. Some features are failed due to this reason.
Such as unable using fcitx in the main program.

This patch remove all libraries. So the netease main program can use libraries
in the system.

Signed-off-by: kXuan <kxuanobj@gmail.com>