Redesign package to use ttf-nerd-fonts-symbols-mono
Patching these fonts based on their originals is a PITA.

It requires installing `font-patcher` from the nerd fonts project.
That's a 2GB tarball, for a package that's just a `makedepends`. It's
not even currently packaged in the AUR or official repos.

I've updated the package to work a bit differently:

- It depends on `ttf-nerd-fonts-symbols-mono` from [community].
- It depends on `otf-operator-mono` which is the unpatched font.
- It'll install a fontconfig rule for the symbols from the
  nerd-fonts-symbols to be preferred when using OperatorMono.

I like the end result a lot better - it relies on a package on
community, doesn't require a huge download to patch things, and fonts
render the same.

The only big difference is that you'll now need to use `OperatorMono` as
a font instead of `OperatorMono Nerd Font`.