Improve version string, remove needless variable use
The prior version was both long and was preceeded by a 'v'.
It has been shortened for readability and the 'v' removed to better follow packaging norms.

The prior version used a _pkgname variable statically declared to terraform instead of just
writing terraform in many places. This has been fixed.