explicitely depend on gtk-sharp-2-git
We need to make sure, that we use a version which is built with libglade 2.x as a dep.
This dep would be pulled by notify-sharp, but gtk-sharp-2 from the repos needed to drop glade as dep, as it moved to only support GTK3.
And without glade-sharp the build will fail.