skip work in post unless needed; use reflink=auto
added a check to the post hook to only copy files back if the original
directory is missing. this stops the hook from copying over the original
when updating a non-running kernel.

switched from `test -e` to `test -d`; -d is preferred for directories

explicitly call `cp --reflink=auto` so we're certain that the hook will
work properly on non reflink capable filesystems. default behavior when
using `--reflink` with the type unspecified is unclear.