set QT_QPA_PLATFORM to xcb
@SamLukeYes reported: when the user set QT_QPA_PLATFORM="wayland;xcb" globally,
the program does not work, because the bundled platform plugins does not
provide native wayland support.

This patch fix this issue by override QT_QPA_PLATFORM to xcb explicitly.

This is a temporary fix. It should be removed when tray menu problem has been
fixed and all bundled libraries removed.

Signed-off-by: kXuan <kxuanobj@gmail.com>