Initial PKGBUILD for osgXR 0.3.0
Create a basic PKGBUILD for osgXR 0.3.0, which uses a PGP signed git tag
to get the release, and CMake to build.

We build the Release build type, but with overridden CXX_FLAGS to avoid
the automatic -O3 while keeping the -DNDEBUG. NDEBUG disables asserts
(which would otherwise introduce absolute source paths into the binary
via __FILE__).

Signed-off-by: James Hogan <james@albanarts.com>