Sort out button number and names mixup
As far as I can tell by documentation (it's a mess!) and testing on my system:
* button 8 = BTN_SIDE = the button that makes Firefox go back
* button 9 = BTN_EXTRA = the button that makes Firefox go forward

However, in the original patches by Kevin it's interpreted like button 8 is
forward and button 9 is back, however the meaning is inverted in both the
spice-gtk and spice-vdagent changes so everything works at the end.

Clean it up to avoid confusion. Also use the SIDE and EXTRA terminology
everywhere since that's what's in spice-protocol for now.