Switch to abootimg 0.6
The abootimg repository on github[1] that is used by the
abootimg-git package doesn't have any tags. In addition
the Guix package for abootimg[2] also uses the source code
provided by Debian to build the abootimg package. So also
using the source code provided by Debian looks like a good
idea.

In both the Debian package and the abootimg-git package
from Aur (which this work is based on) pack-initrd and
unpack-initrd are shipped as abootimg-pack-initrd and
abootimg- unpack-initrd, so we kept that for compatibility
with scripts, instructions, or applications using that.

The Debian patch to abootimg was also added as it at
least adds a manual which is a good thing to have.

The package description was also changed a bit as this tool
it only support the old boot images format.

Note that while the git repository on github[1] doesn't
have new commits since 2012, but having a package for
abootimg (instead of packages for other tools that can
extract components from Android boot images) is a good idea
since abootimg is also used by diffoscope to produce
more fine grained diffs between two Android boot images.

[1]https://github.com/ggrandou/abootimg
[2]https://guix.gnu.org/en/packages/abootimg-0.6/

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>