Use public data.zip available for personal use
It is bit-identical to the one distributed with the game, and makes
building easier.

The rev was bumped because the license files were reworked. The license
in the source code is intended to cover/explain the usage of data.zip
as that is the license referenced from
https://thelettervsixtim.es/makeandplay/. So we let it do that.

But we also include an additional file to try and make clear that this
usage doesn't fall into any of the special allowances for
redistribution. We don't distribute the exceptions because this isn't
one of them, and so could only cause harm by being misunderstood.