Use required Java versions at build and run time
Buck requires Java 8 to build and Java 11 to run. [1]
That has caused spurious failures in the PKGBUILD [2]. The problem is
that the PKGBUILD never enforced (i.e. pinned) the makedepends version
to be actually *used* in the build. So if you have more than one
Java version installed, the PKGBUILD would pick the wrong Java,
causing a build failure. The same goes for the runtime Java version,
where having more than one installed would cause a runtime failure.

Currently, there’s no method to pin JDK/JRE packages to a fixed version
in a vendor-agnostic way [3]. As a workaround, this commit pins the
dependencies to the `-openjdk` vendor, which is better than nothing,
and certainly better than the spurious failures.
The downside is that people who already have JRE 8 or JDK 11 installed
from another vendor will get a useless second copy.

A similar fix has already worked well in the past [4] on macOS. So it’s
pretty safe to assume the fix will be ok here, too.

[1]: https://github.com/facebook/buck/issues/2443

[2]: https://aur.archlinux.org/packages/buck/#comment-738767

[3]: https://wiki.archlinux.org/title/Java#Package_pre-requisites_to_support_archlinux-java

[4]: https://github.com/facebook/homebrew-fb/pull/52

Reported-by: sramey <sramey40@gmail.com>