remove libpng and mesa dependencies
AFAICT libpng was never required - the project uses separate project
stb_image for it's PNG handling.

Originally Mesa was required for the gbm/kms backend. In part due to the
GBM dependency, in part due to the vulkan_intel.h (and respective
extension) used.

The header and the respective extension have been removed with Mesa 21.1

See this upstream report for more:
https://github.com/vkmark/vkmark/issues/33