Update to 1.18.1
There’s no proper GitHub release for this version (only a tag with no
assets), so we have to grab the unversioned zip from CTAN (which will
break with each new release). Also, the contents of the CTAN zip look
quite different, so I just hope this install procedure works.