Initially add veloren-nightly-bin
This package differs from veloren-bin by downloading the precompiled
nightly (instead of precompiled stable) release of Veloren.

This package will update itself, even when not a new version tag is
pushed to the AUR.

This means, that at any given day you can install this package from the
AUR and it will automatically download the binary from that day -
without needing the maintainer to tag a new version in the AUR.

The only downside to this approach is that this will not push updates to
existing users (e.g. yay -Syu), instead they need to manually reinstall
if they want a new version. But even this could be solved by
auto-bumping the version every day with a cronjob (although I kinda like
it this way).