Version 0.18.1
Rewrote the whole PKGBUILD from scratch because a past maintainer had
the /brillant/ idea to erase everything and made a -bin package (it used
to build from sources).

Important: The PKGBUILD builds litecoind and litecoin-cli as intended,
but whether the dependency on Berkely DB 4.8 is necessary is unclear to
me: I only tested that the package builds without errors.