Turn foot-terminfo into an optional dependency
Ncurses 6.3 is coming, and it ships terminfo definitions for
foot.

Foot will default to ncurses’ terminfos. Hence, this patch makes
ncurses a run-time dependency.

However, since they lack certain non-standard capabilites, we still
provide a foot-terminfo package, but with renamed terminfo entries (to
be able to install it side-by-side with ncurses). For this reason,
foot-terminfo is listed as an optional dependency.

To use these terminfo entries, foot must be configured to use
TERM=foot-extra, instead of TERM=foot. This is done by setting
“term=foot-extra” in ~/.config/foot/foot.ini