upgpkg: python-tpm2-pytss-git 0.1.9.r702.f748de7-1: fix running test suite
Since upstream commit
https://github.com/tpm2-software/tpm2-pytss/commit/7d5a0d253a3559ca81d95bc29a582bcbd94e4d6f
the test suite fails because no tests are available in the build/lib.*
directory. Fix this by copying the built library from the build directory to
the source directory and running the test suite from there. This is necessary
as setting PYTHONPATH to build/lib.* as suggested by
https://wiki.archlinux.org/title/Python_package_guidelines#Check does not
appear to work.