Revert to 1.5.3
There is no reason for not updating vmaf, while newer versions offer better
performances and bug fixes. Opening a bug for problems related to
maintenance is not considered to be the official way of doing it, so I'll
need to go through forums and mailing lists. And I don't have the time for
this quest right now (because it is truly a quest).

I had hoped that it would be easier and quicker to have an updated vmaf
package, but it seems that, as many other packages, the last year has been
hard on maintenance and some packages that should have been updated a long
time ago are still lagging behind (as for vmaf).

So I'll just revert the changes introduced by last commit and keep the
updated version on hold.

Signed-off-by: Alexandre Demers <alexandre.f.demers@gmail.com>