remove the pkcs11/ symlink
The token module always looks for libcrypto.so in its own directory
first, followed by /usr/lib next. I'm definitely not going to put the
compat libcrypto.so.1.0.0 symlink in /usr/lib/pkcs11 (as that would
likely affect the newer sac-core and possibly other modules), so the
best option is to keep this old version isolated to its own directory.