bump version to 2.1.2, simpler proxmox dependency patching
Instead of updating the hdozens of places manually, or investing time
into automating that, just use the `[patch.crates-io]` mechanism
cargo provides and clone the main depednecy repos too, which would
have be done anyway by cargo.

Patching out most of the workspace-lib stayed the same.

Also patch our proxmox crate to avoid a glob-import on anyhow, as
that breaks resolbing `Ok()` to te Result tuple, as anyhow added an
Ok helper function in 1.0.49 and glob-imports are *always* wrong..

Signed-off-by: Thomas Lamprecht <thomas@lamprecht.org>