Release 1.0.12
Add '-DCMAKE_CXX_FLAGS=-std=c++11' in order to avoid compilations
issues. For some reason, Sniffles does not update the version string,
than the binary is created inside a folder Sniffles-1.0.11 - instead of
Sniffles-1.0.12. That is why the variable '_pkgver_prev' is used to
hold the previous version and get ride of this bug.