Remove wrapper for starting abaddon
Since abaddon supports the creation of a config file in the users home
directory (more specifically, `pwd`), the highly insecure approach of
saving the configuration in a system-wide accessible (umask 0111)
directory shall be removed.

This commit removes the wrapper script completely and installs abaddon
directly to /usr/bin and removes installing a blank config file with
umask 0111 to /opt/abaddon.

THIS IS A BREAKING CHANGE. Removing the wrapper script (and as a
consequence the cwd in which abaddon is being executed) makes abaddon
unable to find the config file if users don't move the current config
file to their new cwd (i.e. usually their $HOME) and as a consequence
causes settings to be (seemingly) lost. Maybe, a migration mechanism can
be implemented, but would require at least two relbumps.

Signed-off-by: NexAdn <nex@nexadn.de>