5.0.97-beta.59 + update pkgver and source URL
@warwickmm:

> the PKGBUILD is always downloading the latest beta version that
https://www.zotero.org/download/client/dl?channel=beta&platform=linux-x86_64
is serving? Since we are not downloading a specific version, once they
start serving a new version the hashes in the PKGBUILD become out of
date

The true url of the source file includes the git commit hash, e.g.:

https://download.zotero.org/client/beta/5.0.97-beta.59%2Babe8c39c5/Zotero-5.0.97-beta.59%2Babe8c39c5_linux-x86_64.tar.bz2

Updated the $pkgver to include this hash so that the PKGBUILD will point
to a specific file and not be invalidated each time this version
increments. The PKGBUILD will still become out of date on any new
(beta) release but will still build successfully.