Fix re-using old `.jar` files when building
Because the jar file for all versions was downloaded as `Unciv.jar`,
`makepkg` will detect that the file `Unciv.jar` has been downloaded
already and thus not try to re-download it. Thus no new `.jar` file is
downloaded and no actual update is performed. This has been fixed by
naming the `.jar` files according to the version number.